+++ App.js
@@ -1,8 +1,9 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
+import Context from './API/Context';
 import InshortTabs from "./components/InshortTabs"
 
-export default function App() {
+function App() {
   return (
     <View style={{...styles.container, backgroundColor: "#282C35"}}>
       <InshortTabs/>
@@ -16,3 +17,11 @@ const styles = StyleSheet.create({
     marginTop: StatusBar.currentHeight,
   },
 });
+
+export default () => {
+  return(
+    <Context>
+      <App/>
+    </Context>
+  );
+};
