+++ README.md
@@ -1,1 +1,2 @@
 # NewsProject
+# NewsProjectCopy
