+++ components/InshortTabs.js
@@ -1,6 +1,7 @@
-import React, {useState} from 'react'
+import React, {useContext, useState} from 'react'
 import {Text, View, useWindowDimensions} from 'react-native';
 import { SceneMap, TabView } from 'react-native-tab-view';
+import { NewsContext } from '../API/Context';
 import DiscoverScreen from '../Screens/DiscoverScreen';
 import NewsScreen from '../Screens/NewsScreen';
 import TopNavigation from "./TopNavigation"
@@ -8,7 +9,10 @@ import TopNavigation from "./TopNavigation"
 const InshortTabs = () => {
 
     const layout = useWindowDimensions();
-    const [index, setIndex] = useState(1);
+
+    const {index, setIndex} = useContext(NewsContext);
+
+    //const [index, setIndex] = useState(1);
     const [routes] = useState([
         {key: "first", title: "Discover"},
         {key: "second", title: "News"},
