--- components/Screens/DiscoverScreen.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import {Text, View} from 'react-native';
-
-const DiscoverScreen = () => {
-
-    return (
-        <View>
-            <Text>Discover</Text>
-        </View>
-    )
-}
-
-export default DiscoverScreen;
