+++ package.json
@@ -12,12 +12,15 @@
   "dependencies": {
     "@expo/vector-icons": "^13.0.0",
     "@expo/webpack-config": "~0.16.2",
+    "axios": "^0.27.2",
     "expo": "~45.0.0",
     "expo-status-bar": "~1.3.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.1",
+    "react-native-anchor-carousel": "^4.0.1",
     "react-native-pager-view": "5.4.15",
+    "react-native-snap-carousel": "^3.9.1",
     "react-native-tab-view": "^3.1.1",
     "react-native-web": "0.17.7",
     "webpack-dev-server": "~3.11.0"
